+++ src/core/Form/RadioButton/RadioButtonDivider.tsx
@@ -1,10 +1,7 @@
-import React, { Component } from 'react';
+import React, { Component, ReactNode } from 'react';
 import { default as styled } from 'styled-components';
+import { HtmlSpan } from '../../../reset';
 import { TokensProp, InternalTokensProp } from '../../theme';
-import {
-  RadioButtonDivider as CompRadioButtonDivider,
-  RadioButtonDividerProps as CompRadioButtonDividerProps,
-} from '../../../components/Form/RadioButtonDivider';
 import { baseStyles } from './RadioButtonDivider.baseStyles';
 import { withSuomifiDefaultProps } from '../../theme/utils';
 import classnames from 'classnames';
@@ -14,13 +11,15 @@ const radioButtonDividerClassNames = {
   large: `${baseClassName}--large`,
 };
 
-export interface RadioButtonDividerProps
-  extends CompRadioButtonDividerProps,
-    TokensProp {}
+export interface RadioButtonDividerProps extends TokensProp {
+  children: ReactNode;
+  className?: string;
+  variant?: 'small' | 'large';
+}
 
 const StyledRadioButtonDivider = styled(
-  ({ tokens, ...passProps }: RadioButtonDividerProps & InternalTokensProp) => (
-    <CompRadioButtonDivider {...passProps} />
+  ({ className, children }: RadioButtonDividerProps & InternalTokensProp) => (
+    <HtmlSpan className={className}>{children}</HtmlSpan>
   ),
 )`
   ${(props) => baseStyles(props)}
